ipbb/dep2g module rename depparser. Added Changelog and Contributing files. Ready moved to rst format.
