+++ Contributing.rst
@@ -1,0 +1,4 @@
+Contributing
+============
+
+Some help/guidelines if you want to contribute to this library.
