+++ README.rst
@@ -1,6 +1,6 @@
 # IPbus Firmware Builder
 
-### The successor of dep_tree & ProjectManager
+## The successor of dep_tree & ProjectManager
 
 # Getting started
