+++ tox.ini
@@ -1,0 +1,4 @@
+[flake8]
+ignore = F401, E211, E201, E202
+max-line-length = 100
+exclude = tests/*,apidoc/*,conf.py
